# backend_backup/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS School Management API
  description: |
    Complete REST API for the SaaS School Management System.
    
    ## Authentication
    All protected endpoints require JWT authentication via the `Authorization` header.
    Include the tenant ID in the `x-tenant-id` header for multi-tenant isolation.
    
    ## Multi-Tenant Architecture
    The system uses schema-per-tenant isolation. Each school (tenant) has its own
    isolated database schema. All data is automatically scoped to the tenant specified
    in the `x-tenant-id` header.
    
    ## Rate Limiting
    - General API: 100 requests/minute
    - Write Operations: 50 requests/minute
    - Admin Actions: 20 requests/minute
    - Auth Endpoints: 5 requests/minute
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /configuration/branding:
    get:
      summary: Get tenant branding configuration
      responses:
        '200':
          description: Branding config for current tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingConfig'
    put:
      summary: Upsert tenant branding configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandingInput'
      responses:
        '200':
          description: Branding updated
  /configuration/terms:
    get:
      summary: List academic terms
      responses:
        '200':
          description: Academic terms for tenant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcademicTerm'
    post:
      summary: Create or update academic term
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicTermInput'
      responses:
        '201':
          description: Academic term created
  /configuration/classes:
    get:
      summary: List classes
      responses:
        '200':
          description: Classes for tenant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassRecord'
    post:
      summary: Create or update class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassInput'
      responses:
        '201':
          description: Class created
...
  /attendance/report/class:
    get:
      summary: Class attendance report for date
      parameters:
        - name: class_id
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report per status
  /reports/attendance:
    get:
      summary: Aggregate attendance between dates
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: class_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Aggregated attendance rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceAggregate'
  /reports/grades:
    get:
      summary: Grade distribution for exam
      parameters:
        - name: exam_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Grade aggregate rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GradeAggregate'
        '400':
          description: Missing exam id
  /reports/fees:
    get:
      summary: Outstanding fee summary
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, partial, paid, overdue, refunded]
      responses:
        '200':
          description: Fee aggregates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeeAggregate'
    get:
      summary: Get school profile
      responses:
        '200':
          description: School record
    put:
      summary: Upsert school profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchoolInput'
      responses:
        '200':
          description: School updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token in Authorization header
    tenantHeader:
      type: apiKey
      in: header
      name: x-tenant-id
      description: Tenant UUID for multi-tenant isolation
...
  schemas:
...
    BrandingInput:
      type: object
      properties:
        logoUrl:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        themeFlags:
          type: object
          additionalProperties:
            type: boolean
        typography:
          type: object
          additionalProperties: true
        navigation:
          type: object
          additionalProperties: true
    SchoolInput:
      type: object
      properties:
        name:
          type: string
        address:
          type: object
          additionalProperties: true
      required:
        - name
    BrandingConfig:
      type: object
      properties:
        logo_url:
          type: string
          format: uri
        primary_color:
          type: string
        secondary_color:
          type: string
        theme_flags:
          type: object
          additionalProperties:
            type: boolean
        typography:
          type: object
          additionalProperties: true
        navigation:
          type: object
          additionalProperties: true
    AcademicTermInput:
      type: object
      required:
        - name
        - startsOn
        - endsOn
      properties:
        name:
          type: string
        startsOn:
          type: string
          format: date
        endsOn:
          type: string
          format: date
    AcademicTerm:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        starts_on:
          type: string
          format: date
        ends_on:
          type: string
          format: date
    ClassInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    ClassRecord:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
...
    AttendanceAggregate:
      type: object
      properties:
        attendance_date:
          type: string
          format: date
        class_id:
          type: string
          nullable: true
        status:
          type: string
        count:
          type: integer
    GradeAggregate:
      type: object
      properties:
        subject:
          type: string
        grade:
          type: string
        count:
          type: integer
        average_score:
          type: number
          format: float
    FeeAggregate:
      type: object
      properties:
        status:
          type: string
        invoice_count:
          type: integer
        total_amount:
          type: number
          format: float
        total_paid:
          type: number
          format: float
...